extracted identical data and methods from Timeline and Timeline2 and put it into BasicTimeline to generalize both data structures
